+++ src/calculations/__tests__/onResistance.test.ts
@@ -2,15 +2,16 @@ import { readFileSync } from 'fs';
 import { join } from 'path';
 
 import { fromB1505 } from '../../from/b1505';
-import { DispositiveType, onResistance } from '../onResistance';
+import { diodeOnResistance } from '../diodeOnResistance';
+import { transistorOnResistance } from '../transistorOnResistance';
 
 test('Ron transistor', () => {
   const filename = '../../../testFiles/B1505/Output/output.csv';
   const analyses = fromB1505(readFileSync(join(__dirname, filename), 'latin1'));
   for (const analysis of analyses) {
-    const res = onResistance(analysis);
-    expect(res?.value).toBeCloseTo(20.49, 2);
-    expect(res?.score).toBeCloseTo(0.977, 2);
+    const res = transistorOnResistance(analysis);
+    expect(res?.resistanceOn).toBeCloseTo(20.49, 2);
+    expect(res?.score.r2).toBeCloseTo(0.977, 2);
   }
 });
 
@@ -18,9 +19,9 @@ test('Ron diode', () => {
   const filename = '../../../testFiles/B1505/IV/sweep_diode.csv';
   const analyses = fromB1505(readFileSync(join(__dirname, filename), 'latin1'));
   for (const analysis of analyses) {
-    const res = onResistance(analysis, { type: DispositiveType.DIODE });
-    expect(res?.value).toBeCloseTo(20.819, 2);
-    expect(res?.score).toBeCloseTo(0.992, 2);
+    const res = diodeOnResistance(analysis);
+    expect(res?.resistanceOn).toBeCloseTo(20.819, 2);
+    expect(res?.score.r2).toBeCloseTo(0.992, 2);
     expect(res?.Vf).toBeCloseTo(3.05, 2);
     expect(res?.Von).toBeCloseTo(0, 2);
   }
