+++ src/calculations/types.ts
@@ -1,0 +1,11 @@
+import type { RegressionScore } from 'ml-regression-base';
+
+export interface Result {
+  resistanceOn: number;
+  score: RegressionScore;
+  toIndex: number;
+  fromIndex: number;
+}
+export interface ResistanceOptions {
+  delta?: number;
+}
